+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 
 - When signing up with another service (e.g. Google), the platform will now remember a prior language selection
 
-## 2021-08-20
+## 2021-08-26
+
+### Added
+
+- Microsoft Forms integration
 
 ### Fixed